Drastic hack as proof of concept that slightly complicated object will play nice with Google's Protocol Buffers.
I'm working with a friend on his pet project. He's got a set of classes
that he wants to make network transmissible. We're evaluating some
possibilities. I cut down one of the classes a lot to prove that this
will work out pretty well. Some notes to myself in the comments.